requirements.txt
fastapi
uvicorn
transformers
pillow
torch
torchvision
python-multipart
requests
python-dotenv